added middleware

+++ server/src/routes/api/notebook-routes.ts
@@ -1,13 +1,18 @@
 import express from 'express';
-import type { Request, Response } from 'express';
+import type { Response } from 'express';
 import { Notebook } from '../../models/notebook.js';
 import { Note } from '../../models/note.js';
+import { CustomRequest } from '../../middleware/authMiddleware.js';
 
 const router = express.Router();
 
 // GET /api/:id - Get a notebook by ID
 
-router.get('/:id', async (req: Request, res: Response) => {
+router.get('/:id', async (req: CustomRequest, res: Response) => {
+    const reqUser = req.user;
+    if (!reqUser) {
+        return res.status(401).json({ message: 'Not Logged In' });
+    }
     try {
         const notebook = await Notebook.findOne({
             where: { id: req.params.id },
@@ -20,6 +25,9 @@ router.get('/:id', async (req: Request, res: Response) => {
             order: [[Note, 'createdAt', 'ASC']]
         });
         if (notebook) {
+            if (notebook.user_id !== reqUser.id) {
+                return res.status(403).json({ message: 'Unauthorized: ids dont match' });
+            }
             return res.status(200).json({ data: notebook });
         }
         return res.status(404).json({ message: 'Notebook not found' });
@@ -32,11 +40,15 @@ router.get('/:id', async (req: Request, res: Response) => {
 
 // POST /api/notebooks - Create a new notebook
 
-router.post('/', async (req: Request, res: Response) => {
+router.post('/', async (req: CustomRequest, res: Response) => {
+    const reqUser = req.user;
+    if (!reqUser) {
+        return res.status(401).json({ message: 'Not Logged In' });
+    }
     try {
         const notebook = await Notebook.create({
             title: req.body.title,
-            user_id: req.body.user_id,
+            user_id: reqUser.id,
         });
         return res.status(201).json({ data: notebook });
     } catch (err) {
@@ -48,10 +60,17 @@ router.post('/', async (req: Request, res: Response) => {
 
 // PUT /api/notebooks/:id - Update a notebook title by ID
 
-router.put('/:id', async (req: Request, res: Response) => {
+router.put('/:id', async (req: CustomRequest, res: Response) => {
+    const reqUser = req.user;
+    if (!reqUser) {
+        return res.status(401).json({ message: 'Not Logged In' });
+    }
     try {
         const notebook = await Notebook.findByPk(req.params.id);
         if (notebook) {
+            if (notebook.user_id !== reqUser.id) {
+                return res.status(403).json({ message: 'Unauthorized' });
+            }
             notebook.title = req.body.title;
             await notebook.save();
             return res.status(200).json({ data: notebook });
@@ -66,9 +85,14 @@ router.put('/:id', async (req: Request, res: Response) => {
 
 // DELETE /api/notebooks/:id - Delete a notebook by ID
 
-router.delete('/:id', async (req: Request, res: Response) => {
+router.delete('/:id', async (req: CustomRequest, res: Response) => {
+    const reqUser = req.user;
+    if (!reqUser) {
+        return res.status(401).json({ message: 'Not Logged In' });
+    }
+
     try {
-        const deletedNotebook = await Notebook.destroy({ where: { id: req.params.id } });
+        const deletedNotebook = await Notebook.destroy({ where: { id: req.params.id, user_id: reqUser.id } });
         if (deletedNotebook) {
             return res.status(200).json({ message: 'Notebook Deleted', detedNotebook: deletedNotebook });
         }
