+++ server/src/routes/api/user-routes.ts
@@ -1,14 +1,84 @@
 import express from 'express';
-import type { Request, Response } from 'express';
+import type { Response } from 'express';
 import { User } from '../../models/user.js';
 import { Notebook } from '../../models/notebook.js';
 
+import { CustomRequest } from '../../middleware/authMiddleware.js';
+
 const router = express.Router();
+
+// Logged In User Routes
+
+// GET /api/users/me - Get the current user by token
+
+router.get('/me', async (req: CustomRequest, res: Response) => {
+    const reqUser = req.user;
+    if (!reqUser) {
+        return res.status(401).json({ message: 'Unauthorized' });
+    }
+
+    try {
+        const user = await User.findOne({
+            where: { id: reqUser.id },
+            attributes: { exclude: ['password'] },
+            include: [
+                {
+                    model: Notebook,
+                    attributes: ['id', 'title', 'createdAt'],
+                }
+            ],
+            order: [
+                [Notebook, 'createdAt', 'ASC']
+            ]
+        });
+        if (!user) {
+            return res.status(404).json({ message: 'User not found' });
+        }
+        return res.status(200).json({ data: user });
+    } catch (err) {
+        return res.status(400).json({
+            message: (err as Error).message
+        });
+    }
+});
+
+// DELETE /api/users/me - Delete the current user by token
+
+router.delete('/me', async (req: CustomRequest, res: Response) => {
+    const reqUser = req.user;
+    if (!reqUser) {
+        return res.status(401).json({ message: 'Unauthorized' });
+    }
+
+    try {
+        const deletedUser = await User.destroy({ where: { id: reqUser.id } });
+        if (deletedUser) {
+            return res.status(200).json({
+                message: 'User deleted',
+                deletedUser: deletedUser
+            });
+        }
+        return res.status(404).json({ message: 'User not found' });
+    } catch (err) {
+        return res.status(400).json({
+            message: (err as Error).message
+        });
+    }
+})
+
+// Admin (carter) Only Routes - For testing purposes only
 
 // GET /api/users - Get all users
 
-router.get('/', async (_req: Request, res: Response) => {
+router.get('/', async (req: CustomRequest, res: Response) => {
+
+    // So only the user with the username 'carter' can access this route for test purposes
     console.log("GET /api/users");
+    const reqUser = req.user;
+    if (!reqUser || reqUser.username !== 'carter') {
+        return res.status(401).json({ message: 'Unauthorized' });
+    }
+
     try {
         const users = await User.findAll({
             attributes: ['id', 'email', 'username'],
@@ -26,7 +96,15 @@ router.get('/', async (_req: Request, res: Response) => {
 
 // GET /api/users/:id - Get a user by ID
 
-router.get('/:id', async (req: Request, res: Response) => {
+router.get('/:id', async (req: CustomRequest, res: Response) => {
+
+    // So only the user with the username 'carter' can access this route for test purposes
+    console.log("GET /api/users");
+    const reqUser = req.user;
+    if (!reqUser || reqUser.username !== 'carter') {
+        return res.status(401).json({ message: 'Unauthorized' });
+    }
+
     try {
         const user = await User.findOne({
             where: { id: req.params.id },
@@ -53,28 +131,14 @@ router.get('/:id', async (req: Request, res: Response) => {
     }
 });
 
-// POST /api/users - Create a new user
-
-router.post('/', async (req: Request, res: Response) => {
-    console.log("POST /api/users");
-    console.log("Request body: ", req.body);
-    try {
-        const user = await User.create({
-            email: req.body.email,
-            username: req.body.username,
-            password: req.body.password,
-        });
-        return res.status(201).json({ data: user });
-    } catch (err) {
-        return res.status(400).json({
-            message: (err as Error).message
-        });
-    }
-});
-
 // DELETE /api/users - Delete a user by ID
 
-router.delete('/:id', async (req: Request, res: Response) => {
+router.delete('/:id', async (req: CustomRequest, res: Response) => {
+    const reqUser = req.user;
+    if (!reqUser || reqUser.username !== 'carter') {
+        return res.status(401).json({ message: 'Unauthorized' });
+    }
+
     try {
         const deletedUser = await User.destroy({ where: { id: req.params.id } });
         if (deletedUser) {
